Rename SMC files
