+++ pymc/smc/kernels.py
