+++ pymc/smc/sampling.py
@@ -33,7 +33,7 @@ from pymc.backends.arviz import dict_to_dataset, to_inference_data
 from pymc.backends.base import MultiTrace
 from pymc.model import modelcontext
 from pymc.parallel_sampling import _cpu_count
-from pymc.smc.smc import IMH
+from pymc.smc.kernels import IMH
 
 
 def sample_smc(
